+++ galsim/des/des_psfex.py
@@ -81,7 +81,8 @@ class DES_PSFEx(object):
     called an "effective PSF".  Thus, you should not convolve by the pixel profile again
     (nor integrate over the pixel).  This would effectively include the pixel twice!
 
-    @param file_name       The file name to be read in.
+    @param file_name       The file name to be read in, or a pyfits HDU in which case it is used
+                           directly instead of being opened.
     @param image_file_name The name of the fits file of the original image (needed for the
                            WCS information in the header).  If unavailable, you may omit this
                            (or use None), but then the returned profiles will be in image
@@ -90,7 +91,7 @@ class DES_PSFEx(object):
                            image file. (Default `wcs = None`)
     @param dir             Optionally a directory name can be provided if the file_name does not 
                            already include it.  (The image file is assumed to be in the same
-                           directory.) (Default `dir = None`)
+                           directory.) (Default `dir = None`).  Cannot pass an HDU with this option.
     """
     # For config, image_file_name is required, since that always works in world coordinates.
     _req_params = { 'file_name' : str , 'image_file_name' : str }
@@ -102,6 +103,8 @@ class DES_PSFEx(object):
     def __init__(self, file_name, image_file_name=None, wcs=None, dir=None):
 
         if dir:
+            if not isinstance(file_name, basestring):
+                raise ValueError("Cannot provide dir and an HDU instance")
             import os
             file_name = os.path.join(dir,file_name)
             image_file_name = os.path.join(dir,image_file_name)
@@ -118,7 +121,10 @@ class DES_PSFEx(object):
 
     def read(self):
         from galsim import pyfits
-        hdu = pyfits.open(self.file_name)[1]
+        if isinstance(self.file_name, basestring):
+            hdu = pyfits.open(self.file_name)[1]
+        else:
+            hdu = self.file_name
         # Number of parameters used for the interpolation.  We require this to be 2.
         pol_naxis = hdu.header['POLNAXIS']
         if pol_naxis != 2: